Move prerendering config to production config
